pyproject.toml
[tool.poetry]
name = "mqtt-monitor"
version = "0.0.1"
description = "MQTT Monitoring thing"
authors = ["James Hogkinson <james@terminaloutcomes.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
mypy = "^0.931"
black = "^22.1.0"
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
